# Performance improvements to fuzzysearch